Refactoring writesecurely per new typeinject api
