+++ client-src/modules/Gitlab/GitlabAuthentication/GitlabAuthStorage.impl.ts
@@ -1,0 +1,15 @@
+import { GitlabAuthStorage as IGitlabAuthStorage } from "./GitlabAuthStorage.types";
+
+export class GitlabAuthStorage implements IGitlabAuthStorage {
+    getToken() {
+        return window.localStorage.getItem('token');
+    }
+
+    setToken(token: string) {
+        window.localStorage.setItem('token', token);
+    }
+
+    clearToken() {
+        window.localStorage.removeItem('token');
+    }
+}
