+++ client-src/modules/Gitlab/GitlabAuthentication/GitlabAuthStorage.ts
@@ -1,17 +1,7 @@
+import { configureDependency } from "../../../type/inject";
 import { GitlabAuthStorage as IGitlabAuthStorage } from "./GitlabAuthStorage.types";
-import { dependency } from "../../../type/inject";
+import { GitlabAuthStorage } from "./GitlabAuthStorage.impl";
 
-@dependency(IGitlabAuthStorage)
-class GitlabAuthStorage implements IGitlabAuthStorage {
-    getToken() {
-        return window.localStorage.getItem('token');
-    }
-
-    setToken(token: string) {
-        window.localStorage.setItem('token', token);
-    }
-
-    clearToken() {
-        window.localStorage.removeItem('token');
-    }
-}
+configureDependency()
+    .implements(IGitlabAuthStorage)
+    .create(GitlabAuthStorage);
