+++ client-src/modules/QueryBuilder/QueryBuilder.ts
@@ -1,13 +1,7 @@
-import { dependency } from "../../type/inject";
+import { configureDependency } from "../../type/inject";
 import { QueryBuilder as IQueryBuilder } from "./QueryBuilder.types";
-import { parse, stringify } from "query-string";
+import { QueryBuilder } from "./QueryBuilder.impl";
 
-@dependency(IQueryBuilder)
-class QueryBuilder implements IQueryBuilder {
-    getStringFromQuery(query: Object): string {
-        return stringify(query);
-    }
-    getQueryFromString(string: string): Object {
-        return parse(string);
-    }
-}
+configureDependency()
+    .implements(IQueryBuilder)
+    .create(QueryBuilder);
