+++ client-src/type/inject.ts
@@ -1,13 +1,4 @@
-import { createDecorators } from 'typeinject/babel-legacy-decorators';
-import { createContainer, Container } from "typeinject";
+import { createContainer } from "typeinject";
 
-export const container: Container = createContainer();
-export const {
-    get,
-    type,
-    hasCircularDependencies,
-    getImplementation,
-    getImplementations,
-    configure
-} = container;
-export const { dependency, inject } = createDecorators(container);
+export const container = createContainer();
+export const { get, configureDependency, type } = container;
