+++ client-src/ui/components/Editor/prosemirror/Cursor.ts
@@ -1,10 +1,9 @@
 import { EditorPlugins, AddEditorPlugin } from "./EditorPlugins.types";
-import { dependency } from "../../../../type/inject";
 import { Schema } from "prosemirror-model";
 import { dropCursor } from "prosemirror-dropcursor";
 import { gapCursor } from "prosemirror-gapcursor";
 
-@dependency(EditorPlugins)
+
 export class Cursor implements EditorPlugins {
     addEditorPlugins(addEditorPlugin: AddEditorPlugin, schema: Schema) {
         addEditorPlugin(dropCursor());
