+++ client-src/ui/components/Editor/prosemirror/History.ts
@@ -1,11 +1,9 @@
 import { KeyBindings, AddKeyBinding } from "./KeyBindings.types";
-import { dependency, type } from "../../../../type/inject";
 import { Schema } from "prosemirror-model";
 import { undo, redo, history } from "prosemirror-history";
 import { AddEditorPlugin, EditorPlugins } from "./EditorPlugins.types";
 
-@dependency(type(KeyBindings, EditorPlugins))
-class History implements KeyBindings, EditorPlugins {
+export class History implements KeyBindings, EditorPlugins {
     addEditorPlugins(addEditorPlugin: AddEditorPlugin, schema: Schema) {
         addEditorPlugin(history());
     }
