+++ client-src/ui/components/Editor/prosemirror/KeyBindingsManager.ts
@@ -1,4 +1,3 @@
-import { inject, dependency, type } from "../../../../type/inject";
 import { KeyBindings, AddKeyBinding } from "./KeyBindings.types";
 import { KeyBindingsManager as IKeyBindingsManager, KeyMap } from "./KeyBindingsManager.types";
 import { Schema } from "prosemirror-model";
@@ -6,9 +5,7 @@ import { EditorPlugins, AddEditorPlugin } from "./EditorPlugins.types";
 import { keymap } from "prosemirror-keymap";
 import { baseKeymap } from "prosemirror-commands";
 
-@dependency(type(IKeyBindingsManager, EditorPlugins))
-@inject(KeyBindings.multi)
-class KeyBindingsManager implements IKeyBindingsManager, EditorPlugins {
+export class KeyBindingsManager implements IKeyBindingsManager, EditorPlugins {
     private readonly keyBindings: KeyBindings[];
 
     constructor(keyBindings: KeyBindings[]) {
