+++ client-src/ui/components/Editor/prosemirror/marks/EmMark.ts
@@ -1,11 +1,9 @@
 import { EditorMark } from "../EditorMark.types";
-import { dependency, type } from "../../../../../type/inject";
 import { MarkSpec, Schema } from "prosemirror-model";
 import { KeyBindings, AddKeyBinding } from "../KeyBindings.types";
 import { toggleMark } from "prosemirror-commands";
 
-@dependency(type(EditorMark, KeyBindings))
-class LinkMark implements EditorMark, KeyBindings {
+export class EmMark implements EditorMark, KeyBindings {
     name: string = "em";
 
     markSpec: MarkSpec = {
