+++ client-src/ui/components/Editor/prosemirror/nodes/DocNode.ts
@@ -1,9 +1,7 @@
 import { EditorNode } from "../EditorNode.types";
-import { dependency } from "../../../../../type/inject";
 import { NodeSpec } from "prosemirror-model";
 
-@dependency(EditorNode)
-class DocNode implements EditorNode {
+export class DocNode implements EditorNode {
     name: string = "doc";
 
     nodeSpec: NodeSpec = {
