+++ client-src/ui/components/Editor/prosemirror/nodes/HorizontalRuleNode.ts
@@ -1,9 +1,7 @@
 import { EditorNode } from "../EditorNode.types";
-import { dependency } from "../../../../../type/inject";
 import { NodeSpec } from "prosemirror-model";
 
-@dependency(EditorNode)
-class HorizontalRuleNode implements EditorNode {
+export class HorizontalRuleNode implements EditorNode {
     name: string = 'horizontal_rule';
 
     nodeSpec: NodeSpec = {
