+++ client-src/ui/components/Editor/prosemirror/nodes/TextNode.ts
@@ -1,9 +1,7 @@
 import { EditorNode } from "../EditorNode.types";
-import { dependency } from "../../../../../type/inject";
 import { NodeSpec } from "prosemirror-model";
 
-@dependency(EditorNode)
-class TextNode implements EditorNode {
+export class TextNode implements EditorNode {
     name: string = "text";
 
     nodeSpec: NodeSpec = {
